app/requirements.txt
python-telegram-bot==20.3
requests==2.31.0
beautifulsoup4==4.12.2
pytz==2023.3
